Add Django Base

+++ PyDMTQ/PyDMTQ/PyDMTQ.py
@@ -150,23 +150,24 @@ class PyDMTQ(object):
     def saveAllSongData(self,RootPath=os.path.dirname(os.path.abspath(__file__))):
         if not os.path.exists(RootPath):
             os.makedirs(RootPath)
+        SongInfoDict=dict()
         for SongInfo in self.game_getSongList()["songs"]:
             SongID=SongInfo["song_id"]
+            print ("Fetching SongURL of:"+str(SongID))
             URLList=list()
             URLList.extend(self.game_getSongUrl(SongID,"ANDROID","1.0.0")["amazon"])
-            URLList.extend(self.game_getSongUrl(SongID,"IOS","1.0.0")["amazon"])
+            URLList.extend(self.game_getSongUrl(SongID,"IOS","1.0.0")["pmang"])
             for url in URLList:
                 print ("Saving :"+url)
                 SavePath=RootPath+urlparse(url).path
-                if not os.path.exists(os.path.dirname(os.path.abspath(SavePath))):
-                    os.makedirs(os.path.dirname(os.path.abspath(SavePath)))
-                Data=requests.get(url)
-                f=open(SavePath,"w")
-                f.write(Data)
-                f.close
-
+                if os.path.exists(SavePath)==False:
+                    if not os.path.exists(os.path.dirname(os.path.abspath(SavePath))):
+                        os.makedirs(os.path.dirname(os.path.abspath(SavePath)))
+                        Data=requests.get(url)
+                        f=open(SavePath,"w")
+                        f.write(Data)
+                        f.close
 
 
 if __name__ == '__main__':
     x=PyDMTQ("403799106@qq.com","zhs960919")
-    x.saveAllSongData()
